feat(*): include new refactoring code [C2-15]
