+++ app/utils/hashing.py
@@ -1,0 +1,4 @@
+import hashlib
+
+def file_hash(file_bytes: bytes) -> str:
+    return hashlib.sha256(file_bytes).hexdigest()
